% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl7df.R
\name{hl7df}
\alias{hl7df}
\title{Convert a Parsed HL7 Tree to a Data.frame}
\usage{
hl7df(messages)
}
\arguments{
\item{messages}{A parsed HL7 Tree; The result of \code{parsehl7()}}
}
\value{
A Data.frame of HL7 Messages
}
\description{
This function returns a dataframe based on a parsed HL7 tree.
It names columns as the equivalent route to a node.
Values not present in a tree are returned as NA.
An attempt is made to normalize paths between all branches.
}
