% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsehl7.R
\name{parsehl7}
\alias{parsehl7}
\title{A Versatile Parser for HL7 Version 2.x Feeds}
\usage{
parsehl7(feed, file)
}
\arguments{
\item{feed}{A vector representing an HL7 feed, the equivalent output of \code{readLines(file.hl7)}}

\item{file}{A file containing a valid HL7 feed}
}
\value{
A List of Parsed HL7 Messages
}
\description{
This function accepts either a character vector containing separate HL7 segments or a valid HL7 file as input, and returns a parsed R data structure.
}
